BEGINNING NEXT PHASE
I've started re organizing the code so that the final stage of the analysis can be done.
I have done a lot of documentation.
FEC_analyst and Disambiguator are ready to be upgraded.

This is mainly a dummy commit after several months of small changes to the code.
I will branch now to implement a more comprehensive record comparison algorithm using Record objects
and an upgraded Disambiguator.