COMPOUND DISAMBIGUATORS
Disambiguator has a classmethod that combines several D objects into a new one.

To accomplish this, a similar combining classmethod are also implemented in TokenData.

The Affiliation graphs aren't combined. Rather, instead of assigning one affiliation graph to each record, we now assign a list of graphs that will all be used for comparison.

The Disambiguator combiner simply adds all affiliation graphs to a larger list and binds it to the records.