NATIONWIDE
multiprocessing used.

Additional scripts for unit tests created.

The goal is to redesign the code so that one instance can be run for each state, and then the corresponding objects can be combined into a new object for the whole nation.

This means that in many data variables, we must explicitly keep track of the index of the records, and therefore, instead of using lists, we'll have to frequently use dictionaries that map a record index to the target object.

Record.index is now renamed to Record.id

MANY problems with passing project objects between processes.
Some of my classes were not picklable for various reasons:
    -- Tokenizer contained references to functions. Those aren't picklable.
    -- Recursive structures: Each Person used to reference to its neighbor persons directly and this caused recursion limit problems.
    -- Open file objects aren't picklable. I now open and close all file objects explicitly.

A few additional record comparison details implemented. e.g.
    -- If employer is reported as occupation or vice versa, it will be detected.
    -- If first name is entered as initial, it will be handled.

Town class introduced to deal with the Person.neighbor recursion problem.

multiprocessing worker function now strips the project objects of some of its instance objects before sending through pipe to make it serializable:
    project.D.tokenizer.tokenize_functions = None
    project.D.tokenizer.normalize_functions = None
    project.D.project = None

Disambiguator has a method to find all persons with more than one value for a given attribute.
I use this to find persons with more than one STATE in their records.