MAJOR BUGS FOUND!
Some major bugs in affiliation matching where found.

In previous versions, the affiliation graphs were practically ignored! the records need to have a list_G_emoployer field, not just one G_employer. The former is the one that is checked in the comparison function, but when loading the graphs from the file, the latter was being loaded. Now this is fixed.

Another bug was that when checkin for occupation similarity, list_G_employer was being checked!