BUG FIX AND MINOR NEW FEATURES
A serious bug in the output of disambiguation was fixed. The html and json files that output the disambiguate records didn't report the record's id properly. (record data was sorted but the list of ids wasn't similarly sorted). Now this is fixed, and each record comes with its own true id in the html and json output files.

Disambiguator now takes a parameter num_procs for the number of processes to use in update_nearest_neighbors()
disambiguate_main() also has this parameter which it passes along to Disambiguator.

disambiguate_main() also has a whereclause parameter which allows easier customization of the MySQL query run inside it.