WOKIRNG VERSION OF DISTRIBUTED PROCESSING
Here, TestScript.py runs a compound disambiguator which is capable of spawning child processes for
distributed processing of update_nearest_neighbors().

To feed each of the child processes, D.list_of_records and a coupld of other large lists and dicts
are chunked and each chunk is fed to a specific child which then processes the data.
This causes a large increase in memory use which I'm  unhappy with. In effect, the memory used is
doubled in this procedure, because the entire list_of_records is essentially copied (albeit in separate
chunks).

Now, I would like to see if I can used shared memory to avoid this. I'm hoping that the processes
will all be able to see the object D itself, so that there won't be any need to make copies of it
for each child.