DISTRIBUTED, HIGH MEMORY CONSUMPTION
Here I used a "trick" thinking it would improve memory consumption in distributed processing, but was proven wrong.

update_nearest_neighbors() should be able to spawn child processes. These processes each need access to a portion
of the data in D, mainly list_of_records which is a huge list of objects.

In this commit, rather than passing a chunk of the data to the worker functions, I decided to make D a global variable
so that all child processes would have access to it at the same time. This global variable should be defined in the
Disambiguator module so that any process spawned from within this module can see it.

At the same time, I wanted to allow multiple Disambiguators to be able to work simultaneously, which means that all
have to be visible globally but distinguishable from one another, so that each worker child knows exactly which of
the globally available Disambiguator objects spawned it and use only that instance's data. To achieve this, I
wrote a new class ProcessManager which (supposedly) keeps track of all Disambiguator instances created so that
any child worker can retrieve its corresponding D object from the manager by providing an id.

This approach results in a functional algorithm, but fails to save on memory consumption. In fact, it increases
memory consumption dramatically. The reason is that I had misunderstood what's involved in spawning a child process.
As it turns out, each spawned child receives a COPY of the module's namespace, which means everything the ProcessManager
instance contains will be copied as many times as there are processes.

I believe this approach has failed, and I will consider using shared memory objects.