IMPORTANT BUG FIXES
Bug fixed in padded chunking of list_of_records. Previously, chunks did not form a partition of list_of_records. Therefore, upon reconstruction, list_of_records would change.
Now, the process of chunking and reconstructing list_of_records is working as expected.

Also, I finally fixed the initial run of update_nearest_neighbors() in initialize_index_adjacency(). Now, I do this first run by using each record's 'NAME' as its hash. This way, the initial ordering is done according to the names, which gives us the desired blocking.

Tokenizer updated. Addresses are normalized a bit more intelligently. The address parser module screws up po boxes, in which case we simply use the raw string.
Also, a bit more processing is done to extract the middle name. If middle name is a sinle character, we chech if it's in [A-Z]. If there are multiple words in the middle name, we check if there is one that is single-character and choose that. Otherwise, we pick the first word.

update_nearest_neighbors() now handles the single-process and multi-process cases separately by calling different methods for each. e.g., update_nearest_neighbors_multi_proc()