TOKENIZE MULTIPROC & SIGNIFICANCE IN COMPARISON
Now Tokenizer.tokenize() can use multiple processes.
A function in main.py allows us to retrieve records from multiple
states and tokenize all of them at the same time using multiprocessing.

The exported files from this combined tokenization will be saved in
files with "USA" in place of the <state> name.

Also, in Records, the comparison functions don't just return an integer
as the result of the comparison, but a tuple (result_code, other info).
For affiliations comparison functions, we use the second field of the
tuple to return the significance of the compared identifiers as well.