COMPUTING HASHES FOR WHOLE COUNTRY
We can now compute the hashes uniformly for the entire USA using
the combined feature vectors exported to the "USA" files.

Disambiguator has been changed so that the primary product of hash
computing is self.dict_hashes rather than self.LSH_hash. For computing
this dictionary, we no longer need self.list_of_records to be popoulated.
All that is needed is self.dict_vectors.
When self.dict_hashes is computed, it is exported to a file with
prefix "USA".
When a Disambiguator instance needs to actually perform the nearest
neighbor computations, it loads this dict_hashes from file, and from
that, populates self.LSH_hash which is used as before for nearest
neighbor computations.