MAIN OUTSOIRCED, IMPORTS ORGANIZED
All the source code is now in the package disambiguation except for
main.py which is outside. various imports commands are updated to
reflect this and also work with python2.7.3.

Record comparisons are still buggy after changing "result" to tuples.
will be fixed in the next commit.

disambiguate_main whereclause was missing. Corrected. Also, D.tokenizer
was not set. Also fixed.