EDGELIST FROM HASHES
Functions implemented (in hashes.py and main.py) that use the computed
hashes to compute an edgelist of similar records. The results indicate
that with hashes of length 20, it is likely that completely different
records will randomly receive the exact same hash and therefore introduce
errors by appearing identical. Next step is to try longer hashes.