NATIONAL PERSON-LEVEL TOKENDATA COMPUTED
In stage2, before we do combined disambiguation, we first compute
tokendata including token frequencies at the person level, generate
a new TokenData instance with that information and export it to a file
with the "Normalized" label in the file name rather than the Tokenizer
class name. This data will then be used to match records in the final
disambiguation.

The Person class now has a new instance method that computes the dominant
value for a given record attribute amond its records.