NORMALIZED TOKENDATA STAGE2 HAS N_full_name
In stage 2, when we're computing normalized national tokendata, we
keep track of another record attribute, N_full_name as well. This is
because we are interested in the frequency of the full name in additoin
to its parts. The flie USA_Normalized_tokendata will contain full name
frequencies as well.