RECORD COMPARISON CHANGED
name comparison has a new result option: 4, which means the names
are considered identical, but they even have matching middle names.
This is a particularly strong signal and must be differentiated.

occupation comparison has a new result option: 2. (old values of 2
and 3 are converted to 3,4 respectively). 2 means identifiers are
identical, but they are both low-information identifiers such as
RETIRED.

Same thing with employer comparisons.

The verdict functions are updated accordingly. Also, they now check
for the "national" method_id and return results for this option.