TOKENDATA'S NORMALIZED TOKENS ARE USED BY RECORD
Record uses token frequencies for judgments. Previously, it used
tokendat.token_identifiers['LAST_NAME'] to get the code for the
last name tokens. However, it queries for the frequencies of the
*normalized* version of the token, so we add keys for normalized
tokens to tokendata.token_identifiers too, so we can simply query
tokendat.token_identifiers['N_last_name'] instead. This fixes a
bug in stage2 where we initialize the records only with their normalized
tokens. The non-normalized versions of the tokens are only used
for tokenizing them, and not for record comparisons.