RECORD COMPARISONS CHANGED SUBSTANTIALLY.
Record's comparison functions changed (mostly for the "national" method_id)
We also handle the case of identical addresses more carefully. We have for
the first time a verdict function for this case, and it uses affiliation
results as well.
When name comparison result is (2,), we also log the token frequency data
(as we did before with 3,4).