VALIDATE-NATIONAL STARTED
A new script to validate clusters on a national level. This script
selects a random number of target records from the entire database,
then for each target record, it queries the database for all records
with similar names. The target record's identitt, i.e., the "focal
identity" is found, and all its related identities are identified
using the the identity_adjacency of IdentityManager.
The focal identity and its linked identities togeter constitute
the first part of the coding page.
The second part of the coding page consists of all identities
containing any of the "similar" records that were found to have
names similar to the target record.
For each of these two parts an html table is generated and the
two are concatenated into a file.
The clusters in the first part are given an initial selection state
(manifested as a css class) depending on how close they are to the
focal identity. The ones that pass a threshold are given the "score-yes"
class and the others are given the "score-maybeyes" class.