IMPROVED AFFILIATION COMPARISON + SECURITY
A bug is fixed: previously, the possibility that employers and
occupations were mistakenly misreported as one another was not being
actually examined despite the fact that code blocks were written
for the purpose.

The stageII files are properly labeled to differentiate between
ones produced for bootstrapping and ones used for disambiguation
itself. There are separate files for the two now.

The MySQL password is now read from a specified file rather than
being written in the source code.

An encoding problem is solved: employer and occupation strings are
reencoded as ascii upon retrieval from the database now.